+++ railties/test/application/rake_test.rb
@@ -141,12 +141,6 @@ module ApplicationTests
         end
       RUBY
 
-      output = Dir.chdir(app_path){ `bin/rails routes CONTROLLER=cart` }
-      assert_equal ["Passing `CONTROLLER` to `bin/rails routes` is deprecated and will be removed in Rails 5.1.",
-                    "Please use `bin/rails routes -c controller_name` instead.",
-                    "Prefix Verb URI Pattern     Controller#Action",
-                    "  cart GET  /cart(.:format) cart#show\n"].join("\n"), output
-
       output = Dir.chdir(app_path){ `bin/rails routes -c cart` }
       assert_equal "Prefix Verb URI Pattern     Controller#Action\n  cart GET  /cart(.:format) cart#show\n", output
     end